stop getting fake album art from the internet.
A bit a post-mortem on this.
It was causing so many errors, increased initial menu load time by a ton, and at the end of the day was a placeholder to begin with.
We couldn't have shipped with this anyway and reading from an assumed ../albums folder was breaking compiled builds.
I have no idea how this stayed here as long as it did -- probably because it just was never an issue until now.
RIP funny random images, you will be missed, a little bit.