Dragon's cycle test
Has a nice feeling scroll with mouse

Is relative to the center of the screen for better resizing response

Is almost the same equation to actually calc curve, but has better names.

Commented like crazy. More than normal just so it is extra clear.

imperfected pushing up of selected sprite

cannot do an index offset so the selected sprite is always in the middle.

sleep time *yawn*